Update users to have better filenames/paths
