+++ services/app-api/handlers/users/createUser.js
@@ -1,6 +1,6 @@
 import handler from "./../../libs/handler-lib";
 import dynamoDb from "./../../libs/dynamodb-lib";
-import { getUserByUsername } from "./get";
+import { main as getUserByUsername } from "./getUserByUsername";
 
 export const main = handler(async (event, context) => {
   // If this invokation is a prewarm, do nothing and return.
