+++ services/app-api/handlers/users/deleteUser.js
