+++ services/ui-src/src/containers/LocalLogin.js
@@ -14,7 +14,7 @@ export default function Login() {
       firstName: "Alice",
       lastName: "Cooper",
       lastLogin: "01/28/2021",
-      "custom:ismemberof": "state",
+      "custom:ismemberof": "admin",
       // 'custom:ismemberof':
       //   "cn=CARTS_Group_Dev,ou=Groups,dc=cms,dc=hhs,dc=gov,cn=CARTS_Group_Dev_Admin,ou=Groups,dc=cms,dc=hhs,dc=gov,cn=CHIP_D_USER_GROUP_ADMIN,ou=Groups,dc=cms,dc=hhs,dc=gov,cn=IDM_OKTA_TEST,ou=Groups,dc=cms,dc=hhs,dc=gov",
       dateJoined: "01/22/2021",
